+++ src/core/array.js
@@ -1,6 +1,28 @@
 import { createNamespace } from '../core';
 
-export const Array = createNamespace('Array');
+/**
+ * Creates a new wrapped Array chainable.
+ *
+ * @param {any} [arr] - The array to wrap. If a non-array iterable such as a
+ *   string is passed it will be converted to an array with `Array.from`.
+ *   Default is an empty array.
+ *
+ * @returns {SugarChainable<Array>}
+ *
+ * @example
+ *
+ *   new Sugar.Array();
+ *   new Sugar.Array([1,2,3]);
+ *   new Sugar.Array('abc');
+ *
+ **/
+const Namespace = createNamespace('Array', (arg) => {
+  if (arg && !Array.isArray(arg)) {
+    return Array.from(arg);
+  } else {
+    return arg || [];
+  }
+});
 
 export const {
   extend,
@@ -8,4 +30,5 @@ export const {
   defineInstance,
   defineStaticAlias,
   defineInstanceAlias,
-} = Array;
+} = Namespace;
+export { Namespace as Array };