+++ src/core/number.js
@@ -1,6 +1,23 @@
 import { createNamespace } from '../core';
 
-export const Number = createNamespace('Number')
+/**
+ * Creates a new wrapped Number chainable.
+ *
+ * @param {any} [n] - The number to wrap. Non-numbers will be coerced.
+ *   `0` by default.
+ *
+ * @returns {SugarChainable<number>}
+ *
+ * @example
+ *
+ *   new Sugar.Number();
+ *   new Sugar.Number(5);
+ *   new Sugar.Number('5');
+ *
+ **/
+const Namespace = createNamespace('Number', (n = 0) => {
+  return Number(n);
+})
 
 export const {
   extend,
@@ -8,4 +25,5 @@ export const {
   defineInstance,
   defineStaticAlias,
   defineInstanceAlias,
-} = Number;
+} = Namespace;
+export { Namespace as Number };