+++ src/core/regexp.js
@@ -1,6 +1,24 @@
 import { createNamespace } from '../core';
 
-export const RegExp = createNamespace('RegExp');
+/**
+ * Creates a new wrapped RegExp chainable.
+ *
+ * @param {RegExp} reg - The regex to wrap. Will throw an error if not
+ *   provided.
+ *
+ * @returns {SugarChainable<RegExp>}
+ *
+ * @example
+ *
+ *   new Sugar.RegExp(/abc/);
+ *
+ **/
+const Namespace = createNamespace('RegExp', (reg) => {
+  if (!reg) {
+    throw new TypeError('RegExp required');
+  }
+  return reg;
+});
 
 export const {
   extend,
@@ -8,4 +26,5 @@ export const {
   defineInstance,
   defineStaticAlias,
   defineInstanceAlias,
-} = RegExp;
+} = Namespace;
+export { Namespace as RegExp };