+++ src/core/string.js
@@ -1,6 +1,23 @@
 import { createNamespace } from '../core';
 
-export const String = createNamespace('String');
+/**
+ * Creates a new wrapped String chainable.
+ *
+ * @param {any} [n] - The string to wrap. Non-strings will be coerced.
+ *   `0` by default.
+ *
+ * @returns {SugarChainable<string>}
+ *
+ * @example
+ *
+ *   new Sugar.String();
+ *   new Sugar.String('a');
+ *   new Sugar.String(5);
+ *
+ **/
+const Namespace = createNamespace('String', (str = '') => {
+  return String(str);
+});
 
 export const {
   extend,
@@ -8,4 +25,5 @@ export const {
   defineInstance,
   defineStaticAlias,
   defineInstanceAlias,
-} = String;
+} = Namespace;
+export { Namespace as String };